+++ src/py42/services/users.py
@@ -1,7 +1,8 @@
 from py42 import settings
 from py42._compat import quote
-from py42.exceptions import Py42InternalServerError, Py42UserAlreadyExistsError
+from py42.exceptions import Py42InternalServerError, Py42BadRequestError, Py42UserAlreadyExistsError
 from py42.services import BaseService
+from py42.services import handle_active_legal_hold_error
 from py42.services.util import get_all_pages
 
 
@@ -240,7 +241,11 @@ class UserService(BaseService):
         """
         uri = u"/api/UserDeactivation/{}".format(user_id)
         data = {u"blockUser": block_user}
-        return self._connection.put(uri, json=data)
+        try:
+            return self._connection.put(uri, json=data)
+        except Py42BadRequestError as ex:
+            handle_active_legal_hold_error(ex, u"user", user_id)
+            raise
 
     def reactivate(self, user_id, unblock_user=None):
         """Reactivates the user with the given ID.