+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
-## Unreleased
+## 1.9.0 - 2020-10-02
 
 ### Changed
 
@@ -36,6 +36,10 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - Added new exception `Py42UserAlreadyExistsError` to throw if `create_user()` throws `500` and body contains
 `USER_DUPLICATE`
 
+- Added `Py42ActiveLegalHoldError` exception when attempting to deactivate a user or device in an active legal hold.
+    - `py42.sdk.users.deactivate()`
+    - `py42.sdk.devices.deactivate()`
+
 - Added additional user-adjustable setting for security events page size:
     - `py42.settings.security_events_per_page`
 
@@ -58,6 +62,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     - `sdk.auditlogs.get_all()`
 
 ### Changed
+
 - `py42.sdk.queries.query_filter.filter_attributes` renamed to `py42.util.get_attribute_keys_from_class`
 
 ## 1.8.2 - 2020-09-30